datasets better normalization
I really want to figure out how to unify the derives functionality
from the pipelines with the normalization that we do for the tabular
data, however the problem is that it is not possible to invert the
conversion from dict -> list for object values where the inner keys
don't actually have meaning. This is because a _ b and a b _ are
indistinguishable (ab == ab) after the blank is removed.