improvements in the jsonld conversion
Finished implementing context_runtime and now using it to get correct
serialization. Now that we have the apinat-convert script we can get
the flattened jsonld from the nested jsonld, which is something that
pyld can load, for some reason various flattenings produced by the js
version don't work with pyld and some don't work with rdflib jsonld,
all generally unfortunate, but at least we can get the expected
results now. There is still quite a bit of tweaking that will need to
be done to bring the conversion up to parity, but we are not much
closer to being able to ditch the old ttl pipeline.