viewer fix crashing bug on windows
so it turns out it had nothing to do with threads and system* at all,
but instead with a bug in rktio on calls to {directory,file}-exists?
when the dereferenced path was longer than 127 chars, which also
explains the apparent randomness of the occurance, and why I would
rarely trigger it, I was testing datasets with short folder names

for reference see
https://github.com/racket/racket/issues/4233

the workaround we implement is to test using link-exists? which is
safe for our case and does not cause the crash, also reordered the
operations so that the parent path is found without resolving the
symlink on the filesystem so we don't have to worry if for some reason
the link target has a different depth

clean up the old parameterizations of the input and output ports since
they were not actually the problem