viewer xopen-path fix win explorer.exe commas in paths issue
explorer.exe file,name will not work no matter what we do, it will
always fail over to simply opening explorer.exe, therefore we use
set-pwd? and parameterize on the path with the comma, but we then have
to pass path as ".", the double set! is hideous but it works for now