viewer column asc/desc indicator, thread get-jview!
now that the first dataset to appear is not constant there is a
possibility that it might be a dataset with a large export file, which
results in a hang during startup, to avoid this get-jview! now threads
its call to dataset-jview! internally and then checks to see if the ui
is still focused on that dataset when retrieval finishes, this results
in a much smoother ui experience when scrolling through the dataset
list and has generally good ux on unix where the gtk toolkit prevents
us from expanding jviews in the background, overall much better ux for
all platforms if jumping to a new dataset as well since the old jview
is always unparented now so it is clear when the load has completed
and the user can interact

one thing to note is that the way this is set up now the buttons for
the dataset will change before the jview is visible, but that is
probably a good thing, e.g. so that a user can start a fetch+export
while a jview for a large result is still loading