sds 3 changelog rename Consortium data* to Data*
the consortium was overly narrow, there are many possible source of
standards beyond just consortia, such as internal curation standards,
etc. some are checkable some are not, this provides a space to track
those standards and when a validator supports one they recognize it
by name, we'll have to deal with potential collisions, but that can
probably be done by including the standard naming standard (i.e. the
validator implementation) that is assumed, we're not running a DNS
domain name system here